+++ project/src/components/films-list/films-list.tsx
@@ -1,14 +1,14 @@
 import FilmCard from '../../components/film-card/film-card';
 import ShowMoreButton from '../../components/show-more-button/show-more-button';
 import { Film } from '../../types/films';
-import { useState } from 'react';
+import { useState, memo } from 'react';
 
 type FilmsListProps = {
   films: Film[];
   showButton: boolean;
 }
 
-export default function FilmsList({ films, showButton }: FilmsListProps): JSX.Element {
+function FilmsList({ films, showButton }: FilmsListProps): JSX.Element {
   const [activeFilm, setActiveFilm] = useState<number | null>(null);
 
   const filmCardMouseOverHandler = (filmId: number): void => setActiveFilm(filmId);
@@ -33,3 +33,5 @@ export default function FilmsList({ films, showButton }: FilmsListProps): JSX.El
     </div>
   );
 }
+
+export default memo(FilmsList);
